Move hexdump to arch-6502.c
